почти готов бэк!!!

+++ src/Controller/PizzaController.php
@@ -6,6 +6,7 @@ namespace App\Controller;
 
 use App\Entity\Pizza;
 use App\Service\PizzaService;
+use App\Service\SecurityService;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
@@ -15,26 +16,27 @@ use Symfony\Component\Routing\Annotation\Route;
 class PizzaController extends AbstractController
 {
     private PizzaService $pizzaService;
+    private SecurityService $securityService;
 
-    public function __construct(PizzaService $pizzaService)
+    public function __construct(PizzaService $pizzaService, SecurityService $securityService)
     {
         $this->pizzaService = $pizzaService;
+        $this->securityService = $securityService;
     }
 
-    /**
-     * @Route("/pizza", methods={"POST"})
-     */
     public function createPizza(Request $request): JsonResponse
     {
-        $data = json_decode($request->getContent(), true);
+        if (!$this->securityService->isAdmin()) return $this->json(['status' => 'not enough rights'], Response::HTTP_NO_CONTENT);
+
+        $body = json_decode(file_get_contents("php://input"), true);
 
         $pizza = $this->pizzaService->addPizza(
-            $data['pizza_name'],
-            $data['ingredient'],
-            (int)$data['cost']
+            $body['pizza_name'],
+            $body['ingridients'],
+            (int)$body['cost']
         );
 
-        return new JsonResponse(['status' => 'Pizza created', 'id' => $pizza->getPizzaId()], Response::HTTP_CREATED);
+        return $this->json(['status' => 'Pizza created', 'id' => $pizza->getPizzaId()], Response::HTTP_CREATED);
     }
 
     public function listPizzas(): Response
@@ -43,21 +45,22 @@ class PizzaController extends AbstractController
         return $this->json($pizzas, Response::HTTP_OK);
     }
 
-    public function getPizza(int $id): JsonResponse
+    public function getPizza(int $id): Response
     {
         $pizza = $this->pizzaService->getPizza($id);
 
         if (!$pizza) {
-            return new JsonResponse(['status' => 'Pizza not found'], Response::HTTP_NOT_FOUND);
+            return $this->json(['status' => 'Pizza not found'], Response::HTTP_NOT_FOUND);
         }
 
-        return new JsonResponse($pizza, Response::HTTP_OK);
+        return $this->json($pizza, Response::HTTP_OK);
     }
 
-    public function deletePizza(int $id): JsonResponse
+    public function deletePizza(int $id): Response
     {
+        if (!$this->securityService->isAdmin()) return $this->json(['status' => 'not enough rights'], Response::HTTP_OK);
         $this->pizzaService->deletePizza($id);
 
-        return new JsonResponse(['status' => 'Pizza deleted'], Response::HTTP_NO_CONTENT);
+        return $this->json(['status' => 'Pizza deleted'], Response::HTTP_OK);
     }
 }
