+++ src/Controller/UserController.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\Controller;
 
+use App\Service\SecurityService;
 use App\Service\UserService;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
@@ -14,46 +15,49 @@ use Symfony\Component\Routing\Annotation\Route;
 class UserController extends AbstractController
 {
     private UserService $userService;
+    private SecurityService $securityService;
 
-    public function __construct(UserService $userService)
+    public function __construct(UserService $userService, SecurityService $securityService)
     {
         $this->userService = $userService;
+        $this->securityService = $securityService;
     }
 
-    public function createUser(Request $request): Response
+    public function createUser(): Response
     {
+        if (!$this->securityService->isAdmin()) return $this->json(['status' => 'not enough rights'], Response::HTTP_NO_CONTENT);
         $data = json_decode(file_get_contents("php://input"), true);
         $user = $this->userService->addUser(
             $data['user_name'],
-            $data['user_name'],
+            $data['email'],
             $data['password']
         );
-
         return $this->json(['status' => 'User created', 'id' => $user->getUserId()], Response::HTTP_CREATED);
     }
 
     public function listUsers(): Response
     {
+        if (!$this->securityService->isAdmin()) return $this->json(['status' => 'not enough rights'], Response::HTTP_NO_CONTENT);
         $users = $this->userService->listAllUsers();
         
         return $this->json($users, Response::HTTP_OK);
     }
 
-    public function getUserById(int $id): Response
+    public function checkUserName(): Response
     {
-        $user = $this->userService->getUser($id);
+        $data = json_decode(file_get_contents("php://input"), true);
 
-        if (!$user) {
-            return $this->json(['status' => 'User not found'], Response::HTTP_NOT_FOUND);
-        }
-
-        return $this->json($user, Response::HTTP_OK);
+        $name = $data["user_name"];
+        $user = $this->userService->getUser($name);
+        
+        return $this->json(["user" => $user ? $user->toArray() : false], Response::HTTP_OK);
     }
 
     public function deleteUser(int $id): Response
     {
+        if (!$this->securityService->isAdmin()) return $this->json(['status' => 'not enough rights'], Response::HTTP_OK);
         $this->userService->deleteUser($id);
 
-        return $this->json(['status' => 'User deleted'], Response::HTTP_NO_CONTENT);
+        return $this->json(['status' => 'User deleted'], Response::HTTP_OK);
     }
 }
